Branch b_release_1_6_7 Quiet compiler warning in debug message for 32-bit arches. b=18810

+++ lustre/lvfs/fsfilt_ext3.c
@@ -504,8 +504,9 @@ static int fsfilt_ext3_setattr(struct dentry *dentry, void *handle,
         int rc = 0;
 
         LASSERTF(!(iattr->ia_valid & ATTR_SIZE) || !S_ISDIR(inode->i_mode),
-                 "changing i_size on directory #%lu (%p) new %lu old %lu\n",
-                 inode->i_ino, inode, iattr->ia_size, i_size_read(inode));
+                 "changing i_size on directory #%llu (%p) new %llu old %llu\n",
+                 (long long)inode->i_ino, inode, (long long)iattr->ia_size,
+                 (long long)i_size_read(inode));
 
         /* Avoid marking the inode dirty on the superblock list unnecessarily.
          * We are already writing the inode to disk as part of this
